// nextjs-fe/src/features/media/types/file-manager.types.ts
export interface ExtendedFile extends File {
  tempKey?: string;
  isHeavyUploaded?: boolean;
  tempMetadata?: {
    original_name: string;
    extension: string;
    mime_type: string;
    size: number;
  };
}

export interface UploadedFileData {
  file: ExtendedFile | File;
  key: string;
  preview?: string;
  uploading: boolean;
  progress?: number;
  uploaded: boolean;
  error?: string;
  metadata: {
    original_name: string;
    extension: string;
    mime_type: string;
    size: number;
  };
}